requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.2
flask>=2.3.2
Pillow>=10.0.0
gunicorn>=21.2.0